refactor(MtcFeedResource): Move MTC S3 logic to this class.
